Updated Karma, Jsonix, Schema parsers.
Old version of karma no longer would run. Updated to .13 which
works with the current stack and updated the gulpfule to reflect
the new method for starting the server.

Jsonix was also not handling parsing correctly. Updated to the latest
versions from npm.  Removed the bower dependency for Jsonix and the
related schemas.